dist用にJSファイルを修正

+++ js/ts/eventControler.ts
@@ -1,5 +1,5 @@
-import { HandleModal } from "./handleModal";
-import { HandleDropboxFile } from "./handleDropboxFile";
+import { HandleModal } from "./handleModal.js";
+import { HandleDropboxFile } from "./handleDropboxFile.js";
 
 interface _IeventControler {
     setList(listElement: JQuery<HTMLElement>): void;
@@ -13,6 +13,16 @@ interface _IeventControler {
 export class EventControler implements _IeventControler {
     setList(listElement: JQuery<HTMLElement>) {
         // listを表示する
+        const handleDropboxFile = new HandleDropboxFile();
+
+        handleDropboxFile.getFileList().then((listData) => {
+          console.log(listData);
+        },(error) => {
+            alert("erorr");
+        }).catch((error) => {
+            alert("erorr");
+        });
+        console.log(listElement);
     }
 
     reSetList(reSetButtonElement:JQuery<HTMLElement>, listElement: JQuery<HTMLElement>) {
@@ -24,55 +34,68 @@ export class EventControler implements _IeventControler {
             // モーダルウィンドウを表示
             handleModal.showModal();
 
-            handleDropboxFile.getFileList().then(function(listData){
+            handleDropboxFile.getFileList().then((listData) => {
                 // 取得結果をモーダルに表示
                 if(Array.isArray(listData)){
                     handleModal.showResult();
 
+                    const eventControler = new EventControler();
+
+                    // ここが拾ってきたデータの場所
+                    console.log(listData);
+
                     // ファイル一覧を表示
-                    this.setList(listElement);
+                    eventControler.setList(listElement);
 
                 }else{
                     handleModal.showResult("未知のエラー");
                 }
-            },(res) => {
-                handleModal.showResult(res.error);
+            },(error) => {
+                handleModal.showResult(error.error);
             }).catch((e) => {
                 handleModal.showResult(e.message);
             });
         });
     }
 
-    uploadFile(uploadButtonElement: JQuery<HTMLElement>, uploadFileElement: JQuery<HTMLElement>) {
-        uploadButtonElement.on("click",() => {
-            // アップロードファイルが入った時点で読み込みを開始する
-            const handleModal = new HandleModal("アップロード",$(".modal_window"));
+  uploadFile(uploadButtonElement: JQuery<HTMLElement>, uploadFileElement: JQuery<HTMLElement>) {
+    uploadButtonElement.on("click",() => {
+      const filesElement = uploadFileElement.prop('files');
 
-            const handleDropboxFile = new HandleDropboxFile();
+      let uploadFile: File;
 
-            const uploadFile = (<HTMLInputElement>uploadFileElement[0]).files[0];
+      if(filesElement == null){
+        alert("ファイルが登録されていません");
+        return;
+      }else{
+        uploadFile = filesElement[0];
+      }
 
-            try{
-                // モーダルウィンドウを表示
-                handleModal.showModal();
-            }catch{
-                //
-            }
+      const handleModal = new HandleModal("アップロード",$(".modal_window"));
 
-            // アップロード処理実行
-            handleDropboxFile.uploadFile(uploadFile).then((res) => {
-                if(typeof res === "number"){
-                    handleModal.showResult();
-                }else{
-                    handleModal.showResult("未知のエラーが発生しました");
-                }
-            },(res) =>{
-                handleModal.showResult(res.error);
-            }).catch((e) => {
-                handleModal.showResult(e.message);
-            });
-        });
-    }
+      const handleDropboxFile = new HandleDropboxFile();
+
+      try{
+        // モーダルウィンドウを表示
+        handleModal.showModal();
+      }catch{
+        //
+      }
+
+      // アップロード処理実行
+      handleDropboxFile.uploadFile(uploadFile).then((res) => {
+        if(typeof res === "number"){
+          handleModal.showResult();
+        }else{
+          handleModal.showResult("未知のエラーが発生しました");
+        }
+      },(res) =>{
+        handleModal.showResult(res.error);
+      }).catch((e) => {
+        handleModal.showResult(e.message);
+      });
+    });
+  }
 
     downloadFile(downloadLinkElement: JQuery<HTMLElement>) {
         downloadLinkElement.on("click",(event) => {
@@ -128,9 +151,16 @@ export class EventControler implements _IeventControler {
                 // 削除ファイル名一覧を取得
                 deleteFileElementsList = deleteFileElements.find("input[type='checkbox']").filter(":checked");
 
+                deleteFileList = [];
+
                 deleteFileElementsList.each((key, value)=>{
                     deleteFileList.push($(value).data("filename"));
                 });
+
+                if(deleteFileList.length){
+                  alert("1件も選択されていません");
+                  return;
+                }
 
                 // 削除処理実行
                 handleDropboxFile.deleteFiles(deleteFileList).then((res) => {
